api-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic>=2.10.0
easyocr>=1.7.2
Pillow>=11.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiosqlite>=0.21.0
databases>=0.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
greenlet>=3.0.0
